Another bunch of queue deadlock prevention measures
Drop Connect operations when already connected.
isConnected now means that BluetoothGatt object exists in the map AND services were discovered for that BluetoothDevice.
enqueueOperation() is more benevolent.
cancelOperationAfterTimeout() is changed to cancelOperationAndRequeueAfterTimeout(), so the problematic operations is ended and put at the end of the queue again.
cancelOperationAndRequeueAfterTimeout() is applied also for Disconnect, MtuRequest and CharacteristicWrite operations.
MAX_OPS_QUEUE_SIZE is changed to 20.